LFU-194-1 video: rm67191: Add new rm67199 panel support
RM67199 is an new version of this panel and is the same as RM67191
in size, aspect and resolution. The only difference is the DSI
initialization sequence.
We add it support into the RM67191 driver with using different
driver data.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
(cherry picked from commit 2840ef1395692ad9352dc1cb54cd684a99c115d5)