LF-11843: video: imx: Update LVDS PHY settings
Updating the LVDS and Combo PHY settings to match the vendor datasheets
and reference manual.

For both PHYs (iMX8QM and iMX8QXP), the common mode voltage should be 6.

For the LVDS only PHY (iMX8QM), the 'TST' field should be set to the
default value of 0x25. Also, for the LVDS only PHY, the 'M' field should
be set to 1 for the range form 44 MHz to 89 MHz.

For the Combo PHY (iMX8QXP), the 'CO' field need to be set to 1 to select
a PLL output divide of 2.

Signed-off-by: Oliver F. Brown <oliver.brown@oss.nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>